MQTT/requirements.txt
paho-mqtt==1.6.1
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
